requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
joblib==1.5.3
numpy==2.2.6
packaging==26.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
